Move helpers `hasNext()` and `hasComponents()` into model
