+++ designer/client/src/components/Menu/Menu.tsx
@@ -1,4 +1,4 @@
-import { hasListField } from '@defra/forms-model'
+import { hasComponents, hasListField } from '@defra/forms-model'
 import { highlightAll } from 'prismjs'
 import React, { useContext } from 'react'
 
@@ -13,7 +13,6 @@ import { RenderInPortal } from '~/src/components/RenderInPortal/RenderInPortal.j
 import { Tabs } from '~/src/components/Tabs/Tabs.jsx'
 import { ConditionsEdit } from '~/src/conditions/ConditionsEdit.jsx'
 import { DataContext } from '~/src/context/DataContext.js'
-import { hasComponents } from '~/src/data/definition/hasComponents.js'
 import { i18n } from '~/src/i18n/i18n.jsx'
 import { ListsEdit } from '~/src/list/ListsEdit.jsx'
 import { ListsEditorContextProvider } from '~/src/reducers/list/listsEditorReducer.jsx'
