+++ designer/client/src/components/Page/Page.tsx
@@ -1,4 +1,5 @@
 import {
+  hasComponents,
   slugify,
   type ComponentDef,
   type Page as PageType
@@ -11,7 +12,6 @@ import { ComponentCreate } from '~/src/components/ComponentCreate/ComponentCreat
 import { Flyout } from '~/src/components/Flyout/Flyout.jsx'
 import { RenderInPortal } from '~/src/components/RenderInPortal/RenderInPortal.jsx'
 import { DataContext } from '~/src/context/DataContext.js'
-import { hasComponents } from '~/src/data/definition/hasComponents.js'
 import { findSection } from '~/src/data/section/findSection.js'
 import { i18n } from '~/src/i18n/i18n.jsx'
 import { ComponentContextProvider } from '~/src/reducers/component/componentReducer.jsx'
