+++ designer/client/src/data/condition/removeCondition.ts
@@ -1,6 +1,8 @@
-import { type ConditionWrapper, type FormDefinition } from '@defra/forms-model'
-
-import { hasNext } from '~/src/data/page/hasNext.js'
+import {
+  hasNext,
+  type ConditionWrapper,
+  type FormDefinition
+} from '@defra/forms-model'
 
 /**
  * Remove link condition
