+++ designer/client/src/data/page/deleteLink.ts
@@ -1,8 +1,7 @@
-import { type FormDefinition, type Page } from '@defra/forms-model'
+import { hasNext, type FormDefinition, type Page } from '@defra/forms-model'
 
 import { findLink } from '~/src/data/page/findLink.js'
 import { findPage } from '~/src/data/page/findPage.js'
-import { hasNext } from '~/src/data/page/hasNext.js'
 
 /**
  * Delete link from page
